// src/api/generated/models/apiclan-top.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Authentication
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface APIClanTop
 */
export interface APIClanTop {
    /**
     * 
     * @type {string}
     * @memberof APIClanTop
     */
    'region': string;
    /**
     * 
     * @type {string}
     * @memberof APIClanTop
     */
    'name': string;
    /**
     * 
     * @type {number}
     * @memberof APIClanTop
     */
    'clan_id': number;
    /**
     * 
     * @type {string}
     * @memberof APIClanTop
     */
    'tag': string;
    /**
     * 
     * @type {number}
     * @memberof APIClanTop
     */
    'general_battles': number;
    /**
     * 
     * @type {number}
     * @memberof APIClanTop
     */
    'general_wins': number;
    /**
     * 
     * @type {number}
     * @memberof APIClanTop
     */
    'averageDamage': number;
    /**
     * 
     * @type {number}
     * @memberof APIClanTop
     */
    'rating': number;
}